Tighten up queue implementation
* Add a test suite for queues.
* Unpack the queue digits into the nodes to reduce time,
  residency, allocation, and indirection.
* Coax GHC to produce a proper worker-wrapper pair for
  `Array.append`. I don't know why this doesn't just work
  out of the box.
