+++ test/Stack.hs
